# specs/001-todo-full-stack-app/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Todo Full-Stack Web Application API
  version: 1.0.0
  description: API for managing user tasks in the Todo application
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.yourapp.com
    description: Production server
paths:
  /api/{user_id}/tasks:
    get:
      summary: Get all tasks for a user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the user
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
          description: Number of tasks to return
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Number of tasks to skip
        - name: completed
          in: query
          schema:
            type: boolean
          description: Filter by completion status
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: User not found
      security:
        - bearerAuth: []
    post:
      summary: Create a new task for a user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid request body
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation error
      security:
        - bearerAuth: []
  /api/{user_id}/tasks/{id}:
    get:
      summary: Get a specific task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the user
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the task
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Task or user not found
      security:
        - bearerAuth: []
    put:
      summary: Update a task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the user
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid request body
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Task or user not found
        '422':
          description: Validation error
      security:
        - bearerAuth: []
    delete:
      summary: Delete a task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the user
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the task
      responses:
        '204':
          description: Task deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Task or user not found
      security:
        - bearerAuth: []
  /api/{user_id}/tasks/{id}/complete:
    patch:
      summary: Update task completion status
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the user
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                completed:
                  type: boolean
              required:
                - completed
      responses:
        '200':
          description: Task completion status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid request body
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Task or user not found
        '422':
          description: Validation error
      security:
        - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 255
        description:
          type: string
          nullable: true
        completed:
          type: boolean
        due_date:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        user_id:
          type: string
          format: uuid
      required:
        - id
        - title
        - completed
        - created_at
        - updated_at
        - user_id
    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
          nullable: true
        due_date:
          type: string
          format: date-time
          nullable: true
      required:
        - title
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
          nullable: true
        completed:
          type: boolean
        due_date:
          type: string
          format: date-time
          nullable: true